requirements.txt
# ==============================================
# CORE DEPENDENCIES
# ==============================================
fastapi==0.104.1
celery==5.3.4
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.13.1

# ==============================================
# DOCUMENT PROCESSING
# ==============================================
mammoth==1.6.0
python-docx==0.8.11
beautifulsoup4==4.12.2
pandoc==2.3

# ==============================================
# ML/NLP - VERSIONES COMPATIBLES PYTHON 3.12
# ==============================================
sentence-transformers==2.2.2
scikit-learn==1.4.0
numpy==1.26.4
pandas==2.1.4
nltk==3.8.1

# ==============================================
# EXTERNAL APIS
# ==============================================
anthropic==0.7.8
deepl==1.16.1
requests==2.31.0
aiohttp==3.9.1
aiofiles==23.2.0

# ==============================================
# WEB AND ASYNC
# ==============================================
uvicorn[standard]==0.24.0
streamlit==1.28.1

# ==============================================
# UTILITIES
# ==============================================
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
GitPython==3.1.40
plotly==5.17.0
openpyxl==3.1.2
jinja2==3.1.2

# ==============================================
# TESTING
# ==============================================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# ==============================================
# DEVELOPMENT TOOLS
# ==============================================
black==23.11.0
flake8==6.1.0
mypy==1.7.1